supernova refactor test to separate mod
